# doc/api.yaml
openapi:  3.0.3
info: 
  title: Wasa Photo - OpenAPI
  description: |
    API for sharing photos.
    
    Each user will be presented with a stream of photos (images) in reverse chronological order, with information about when each photo was uploaded (date and time) and how many likes and comments it has. The stream is composed by photos from “following” (other users that the user follows). Users can place (and later remove) a “like” to photos from other users. Also, users can add comments to any image (even those uploaded by themself). Only authors can remove their comments. Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about Alice. Alice can decide to remove the ban at any moment. Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse chronological order), how many photos have been uploaded, and the user’s followers and following. Users can change their usernames, upload photos, remove photos, and follow/unfollow other users. Removal of an image will also remove likes and comments.
      A user can search other user profiles via username
  version: 1.0.1
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    username:
      schema:
        description: username schema.
        type: string
        pattern: '^.*?$'
        example: SiweiHu
        minLength: 5
        maxLength: 16
      name: username
      in: path
      required: true
      description: identifier for users
...
paths:
...
  /users/{username}:
    parameters:
       - $ref: "#/components/parameters/username"

    put:
      security:
        - bearerAuth: []
      tags:
        - user
      operationId: setMyUserName
      summary: set the username specified by the userin the request
      requestBody:
        content:
          application/json:
            schema: 
                type: string
                example: SiweiHu
                pattern: '^.*?$'
                minLength: 3
                maxLength: 16
      responses:
        "201":
          description: Username updated successfully
          content:
            application/json:
              schema: 
                type: string
                example: SiweiHu
                pattern: '^.*?$'
                minLength: 3
                maxLength: 16
        "400": 
          description: The request was not compliant with the documentation (eg. missing fields, etc).
        "500": 
          description: The server encountered an internal error. Further info in server logs.
...